Move the config path helpers next to the state that uses them
update_data_paths had exactly one caller, AppState.from_config, and
get_expanded_path had exactly one caller in turn. They were the first half of
normalising a config rather than utilities the app reaches for, so helpers/ was
holding state's private working. It is now what its name suggests from the
outside: the fzf boundary, and nothing else.

The two tests follow the function into test_state.
